+++ client/src/pro/index.js
@@ -3,4 +3,5 @@ export default {
     GroupsTab: null,
     RouteLBSettings: null,
     ConstellationNATSReplicas: null,
+    DeploymentForm: null
 }