+++ src/httpServer.go
@@ -700,6 +700,8 @@ func InitServer() *mux.Router {
 	srapiAdmin.HandleFunc("/api/constellation/tunnels", constellation.TunnelList)
 	srapiAdmin.HandleFunc("/api/constellation/edit-device", constellation.DeviceEdit_API)
 	srapiAdmin.HandleFunc("/api/constellation/get-next-ip", constellation.API_GetNextIP)
+	srapiAdmin.HandleFunc("/api/constellation/deployments", constellation.DeploymentsRoute)
+	srapiAdmin.HandleFunc("/api/constellation/deployments/{name}", constellation.DeploymentsIdRoute)
 
 	srapiAdmin.HandleFunc("/api/events", metrics.API_ListEvents)